Backend/requirements.txt
# ========================================
# TRIP EXPLORER - Backend Requirements
# ========================================
# Python 3.13+
# Django REST API avec intégration TripAdvisor

# Core Django Framework
# ========================================
Django==5.2.7
djangorestframework==3.15.2

# Authentication & Security
# ========================================
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# CORS & HTTP
# ========================================
# Gère les requêtes cross-origin pour le frontend React
django-cors-headers==4.6.0

# HTTP requests pour l'API TripAdvisor
requests==2.32.4
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11

# Configuration & Environment
# ========================================
# Gestion des variables d'environnement (.env)
python-dotenv==1.1.1
python-decouple==3.8

# Database & ORM
# ========================================
# SQLite inclus avec Django (production: PostgreSQL recommandé)
sqlparse==0.5.3

# Django Extensions & Utilities
# ========================================
django-extensions==3.2.3
django-filter==24.3

# ASGI/WSGI Server Support
# ========================================
asgiref==3.10.0

# Timezone Support
# ========================================
tzdata==2025.2

# Development & Testing (optionnel)
# ========================================
# pytest==8.3.5
# pytest-django==4.9.0
# coverage==7.6.9

# Production Server (optionnel)
# ========================================
# gunicorn==23.0.0
# psycopg2-binary==2.9.10  # Pour PostgreSQL